Do not crash anymore when encountering a non-existing enum value in a file that is read with the input framework.
Instead, output an error message.

What is a little bit... well, not exactly nice about the way this is
currently performed is, that the current line is not skipped. Instead, the
value will simply be set to 0. This does not really lead to problems, but
(depending on the setting) might lead to other errors being reported later
when this value is being used somewhere.

This is due to the structure of the input framework, at the moment there is
no easy way to just skip processing of the entire line. That might be nice,
but it looks that it would involve touching quite a bit of code...

Addresses BIT-1229